Fixed bug in performance API that caused PerfApi::drainStore to sometimes drain to the wrong address: The addr parameter in PerfApi::commitMemoryWrite was of type unsigned which led to loss of address bits. The type is now uint64_t.

+++ PerfApi.cpp
@@ -806,7 +806,7 @@ PerfApi::setStoreData(unsigned hartIx, uint64_t tag, uint64_t value)
 
 
 bool
-PerfApi::commitMemoryWrite(Hart64& hart, unsigned addr, unsigned size, uint64_t value)
+PerfApi::commitMemoryWrite(Hart64& hart, uint64_t addr, unsigned size, uint64_t value)
 {
   switch (size)
     {
